Move "deferred" utils to async module.
