+++ src/client/common/net/socket/socketServer.ts
@@ -1,8 +1,8 @@
 import { EventEmitter } from 'events';
 import { injectable } from 'inversify';
 import * as net from 'net';
-import { createDeferred, Deferred } from '../../helpers';
 import { ISocketServer } from '../../types';
+import { createDeferred, Deferred } from '../../utils/async';
 import { noop } from '../../utils/misc';
 
 @injectable()
