+++ src/client/debugger/DebugServers/BaseDebugServer.ts
@@ -1,11 +1,11 @@
 // tslint:disable:quotemark ordered-imports no-any no-empty
-"use strict";
+'use strict';
 
-import { DebugSession } from "vscode-debugadapter";
-import { IPythonProcess, IDebugServer } from "../Common/Contracts";
-import { EventEmitter } from "events";
-import { Deferred, createDeferred } from '../../common/helpers';
+import { DebugSession } from 'vscode-debugadapter';
+import { IPythonProcess, IDebugServer } from '../Common/Contracts';
+import { EventEmitter } from 'events';
 import { Socket } from 'net';
+import { Deferred, createDeferred } from '../../common/utils/async';
 
 export abstract class BaseDebugServer extends EventEmitter {
     protected clientSocket: Deferred<Socket>;
@@ -15,8 +15,11 @@ export abstract class BaseDebugServer extends EventEmitter {
     protected pythonProcess: IPythonProcess;
     protected debugSession: DebugSession;
 
-    protected isRunning: boolean;
+    protected isRunning?: boolean;
     public get IsRunning(): boolean {
+        if (this.isRunning === undefined) {
+            return false;
+        }
         return this.isRunning;
     }
     protected debugClientConnected: Deferred<boolean>;
