+++ src/client/interpreter/locators/services/cacheableLocatorService.ts
@@ -7,8 +7,8 @@ import { injectable, unmanaged } from 'inversify';
 import * as md5 from 'md5';
 import { Uri } from 'vscode';
 import { IWorkspaceService } from '../../../common/application/types';
-import { createDeferred, Deferred } from '../../../common/helpers';
 import { IPersistentStateFactory } from '../../../common/types';
+import { createDeferred, Deferred } from '../../../common/utils/async';
 import { IServiceContainer } from '../../../ioc/types';
 import { IInterpreterLocatorService, PythonInterpreter } from '../../contracts';
