+++ src/test/common/variables/envVarsProvider.multiroot.test.ts
@@ -7,10 +7,10 @@ import * as fs from 'fs-extra';
 import { EOL } from 'os';
 import * as path from 'path';
 import { ConfigurationTarget, Disposable, Uri, workspace } from 'vscode';
-import { createDeferred } from '../../../client/common/helpers';
 import { NON_WINDOWS_PATH_VARIABLE_NAME, WINDOWS_PATH_VARIABLE_NAME } from '../../../client/common/platform/constants';
 import { IDisposableRegistry, IPathUtils, IsWindows } from '../../../client/common/types';
 import { IS_WINDOWS } from '../../../client/common/util';
+import { createDeferred } from '../../../client/common/utils/async';
 import { EnvironmentVariablesService } from '../../../client/common/variables/environment';
 import { EnvironmentVariablesProvider } from '../../../client/common/variables/environmentVariablesProvider';
 import { EnvironmentVariables } from '../../../client/common/variables/types';
