+++ src/test/linters/lint.provider.test.ts
@@ -5,9 +5,9 @@ import { Container } from 'inversify';
 import * as TypeMoq from 'typemoq';
 import * as vscode from 'vscode';
 import { IDocumentManager } from '../../client/common/application/types';
-import { createDeferred } from '../../client/common/helpers';
 import { IFileSystem } from '../../client/common/platform/types';
 import { IConfigurationService, ILintingSettings, IPythonSettings, Product } from '../../client/common/types';
+import { createDeferred } from '../../client/common/utils/async';
 import { IInterpreterService } from '../../client/interpreter/contracts';
 import { ServiceContainer } from '../../client/ioc/container';
 import { ServiceManager } from '../../client/ioc/serviceManager';
