+++ src/test/unittests/display/main.test.ts
@@ -10,8 +10,8 @@ import * as typeMoq from 'typemoq';
 import { StatusBarItem, Uri } from 'vscode';
 import { IApplicationShell } from '../../../client/common/application/types';
 import { Commands } from '../../../client/common/constants';
-import { createDeferred } from '../../../client/common/helpers';
 import { IConfigurationService, IPythonSettings, IUnitTestSettings } from '../../../client/common/types';
+import { createDeferred } from '../../../client/common/utils/async';
 import { noop } from '../../../client/common/utils/misc';
 import { IServiceContainer } from '../../../client/ioc/types';
 import { CANCELLATION_REASON } from '../../../client/unittests/common/constants';
