+++ src/test/unittests/stoppingDiscoverAndTest.test.ts
@@ -5,8 +5,8 @@ import { expect, use } from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import * as path from 'path';
 import { Uri } from 'vscode';
-import {createDeferred} from '../../client/common/helpers';
 import { Product } from '../../client/common/types';
+import { createDeferred } from '../../client/common/utils/async';
 import { CANCELLATION_REASON, CommandSource, UNITTEST_PROVIDER } from '../../client/unittests/common/constants';
 import { ITestDiscoveryService } from '../../client/unittests/common/types';
 import { initialize, initializeTest } from '../initialize';
